+++ app/Models/User.php
@@ -7,6 +7,10 @@ use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
 use Spatie\Permission\Traits\HasRoles;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
+
+use Illuminate\Database\Eloquent\Relations\BelongsToMany;
+
 class User extends Authenticatable
 {
     /** @use HasFactory<\Database\Factories\UserFactory> */
@@ -21,6 +25,10 @@ class User extends Authenticatable
         'name',
         'email',
         'password',
+        'city_id',
+        'province_id',
+        'district_id',
+        'postal_code',
     ];
 
     /**
@@ -44,5 +52,37 @@ class User extends Authenticatable
             'email_verified_at' => 'datetime',
             'password' => 'hashed',
         ];
+    }
+
+    public function city(): BelongsTo
+    {
+        return $this->belongsTo(City::class, 'city_id', 'city_id');
+    }
+
+    public function province(): BelongsTo
+    {
+        return $this->belongsTo(Province::class, 'province_id', 'prov_id');
+    }
+
+    public function district(): BelongsTo
+    {
+        return $this->belongsTo(District::class, 'district_id', 'dis_id');
+    }
+
+    
+
+    public function transactionCategories(): BelongsToMany
+    {
+        return $this->belongsToMany(TransactionCategory::class);
+    }
+
+    public function mailCategories(): BelongsToMany
+    {
+        return $this->belongsToMany(MailCategory::class);
+    }
+
+    public function transactionTypes(): BelongsToMany
+    {
+        return $this->belongsToMany(TransactionType::class);
     }
 }